my schedule added; mock data moved
