--- frontend/src/components/MySchedule.js
@@ -1,12 +1,0 @@
-import React from "react";
-import Grid from "./Grid";
-
-const MySchedule = () => {
-    return (
-        <div>
-            <Grid />
-        </div>
-    )
-}
-
-export default MySchedule;
